feat: payment checkout in separated component
